feat(tui): #275 split built-ins into static actions + dynamic sources
